circular dependency and unnecessary lines removed
