+++ src/encoded/static/components/genome_browser.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import _ from 'underscore';
 import { FetchedData, Param } from './fetched';
 import AutocompleteBox from './region_search';
-import { filterForVisualizableFiles } from './filegallery';
 
 const domainName = 'https://www.encodeproject.org';
 
@@ -46,6 +45,22 @@ const dummyFiles = [
         href: '/files/ENCFF847CBY/@@download/ENCFF847CBY.bigBed',
     },
 ];
+
+// Not all files can be visualized on the Valis genome browser
+// Some of these files should be visualizable later, after updates to browser
+export function filterForVisualizableFiles(fileList) {
+    const newFileList = fileList.filter(file => (
+        (file.file_format === 'bigWig' || file.file_format === 'bigBed')
+        && (file.file_format_type !== 'bedMethyl')
+        && (file.file_format_type !== 'bedLogR')
+        && (file.file_format_type !== 'idr_peak')
+        && (file.file_format_type !== 'tss_peak')
+        && (file.file_format_type !== 'pepMap')
+        && (file.file_format_type !== 'modPepMap')
+        && ['released', 'in progress', 'archived'].indexOf(file.status) > -1
+    ));
+    return newFileList;
+}
 
 // Fetch gene coordinate file
 export function getCoordinateData(geneLink, fetch) {
@@ -134,8 +149,7 @@ class GenomeBrowser extends React.Component {
                 domain = domainName;
                 newFiles = [...this.state.pinnedFiles, ...dummyFiles];
             } else {
-                let propsFiles = this.props.files;
-                propsFiles = filterForVisualizableFiles(propsFiles);
+                const propsFiles = filterForVisualizableFiles(this.props.files);
                 files = _.chain(propsFiles)
                     .sortBy(obj => obj.output_type)
                     .sortBy((obj) => {
@@ -281,8 +295,7 @@ class GenomeBrowser extends React.Component {
             newFiles = [...this.state.pinnedFiles, ...dummyFiles];
         } else {
             // Filter files to include only bigWig and bigBed formats, and not 'bigBed bedMethyl' formats and only released or in progress files
-            let propsFiles = this.props.files;
-            propsFiles = filterForVisualizableFiles(propsFiles);
+            const propsFiles = filterForVisualizableFiles(this.props.files);
             // Set default ordering of tracks to be first by replicate then by output_type
             // Ordering by replicate is like this: 'Rep 1,2' -> 'Rep 1,3,...' -> 'Rep 2,3,...' -> 'Rep 1' -> 'Rep 2' -> 'Rep N'
             // Multiplication by 1000 orders the replicates with a single replicate at the end
